+++ src/main/java/org/summerboot/jexpress/security/SecurityUtil.java
@@ -23,9 +23,9 @@ import org.apache.commons.lang3.RegExUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.text.StringEscapeUtils;
 import org.summerboot.jexpress.boot.BootErrorCode;
-import org.summerboot.jexpress.nio.server.SessionContext;
-import org.summerboot.jexpress.nio.server.domain.Err;
+import org.summerboot.jexpress.controller.Err;
 import org.summerboot.jexpress.util.ApplicationUtil;
+import org.summerboot.jexpress.controller.SessionContext;
 
 import javax.net.ssl.HostnameVerifier;
 import javax.net.ssl.HttpsURLConnection;