+++ CHANGES.md
@@ -245,7 +245,7 @@ New configuration items have been added to set the server-specific idle connecti
     * `@Controller.implTag` renamed to `@Controller.AlternativeName`.
     * `@Service.implTag` renamed to with `@Service.AlternativeName`.
     * `@Log.hideJsonStringFields`, `hideJsonNumberFields` and `hideJsonArrayFields` all renamed to with `@Log.maskDataFields`.
-    * `org.summerboot.jexpress.nio.server.domain.ServiceContext` renamed to `org.summerboot.jexpress.nio.server.SessionContext`.
+    * `org.summerboot.jexpress.nio.server.domain.ServiceContext` renamed to `org.summerboot.jexpress.controller.SessionContext`.
     * `@ImportResource.checkImplTagUsed` renamed to `@ImportResource.whenUseAlternative`.
     * `@ImportResource.loadWhenImplTagUsed` renamed to `@ImportResource.thenLoadConfig`.
 
@@ -464,13 +464,13 @@ New configuration items have been added to set the server-specific idle connecti
 * Other than `j_security_check`, also supported `POST ${context-root}/login` with JSON request in `BootController {"username" : "value", "password" : "value"}`.
 * Let the controller layer handle LDAP escape.
 * Logged both raw and decoded uri.
-* Refactored lifecycle/exception listeners/handlers to `org.summerboot.jexpress.boot.event`.
+* Refactored lifecycle/exception listeners/handlers to `org.summerboot.jexpress.boot.lifecycle`.
 
 ## Version 2.4.1
 
 * Enabled `@jakarta.validation.constraints.Pattern` for RESTFul api parameters: `PathParam`, `MatrixParam`, `QueryParam`, `FormParam`, `HeaderParam`, and `CookieParam`.
 * Fixed: cli `-decrypt` generates `cfg_grpc.properties` even if there is no gRPC impl.
-* Log trace enable on `org.summerboot.jexpress.nio.server.BootHttpRequestHandler` will override `@org.summerboot.jexpress.boot.annotation.Log` settings to log all requests and responses.
+* Log trace enable on `org.summerboot.jexpress.controller.restful.BootHttpRequestHandler` will override `@org.summerboot.jexpress.boot.annotation.Log` settings to log all requests and responses.
 * Reformatted Java source code via IDEA default formatter.
 
 ## Version 2.3.13